+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 - New pelican by ENDYLAU
+- Iconic Framework support
+- Launcher Drawer support
 
 ### Fixed
 - Issue where front texture info is lost